# This is a combination of 2 commits. # The first commit's message is:
Added support for altering table type and runtime helpers for this. Here, we start to grapple with the fact that we can optimize how table copying is done, as well as merge other tasks (such as layout transform) into the same task. Looks like that will make migration planning more complex when applicable, but that the final result should be faster. We should still detect when copying is NOT needed, and handle those cases less expensively. Also, we should probably produce a correct-and-complete-but-maybe-slow version before starting in on the complex optimizations.

# This is the 2nd commit message:

Added test for whether table migration requires copying. This will be used to decide migration strategy.